[feature/small-improvements - Change int to float for temp in DailyForecastDTO

+++ bop_common/dtos/weather/daily_forecast_dto.py
@@ -3,5 +3,5 @@ from dataclasses import dataclass
 @dataclass
 class DailyForecastDTO:
     date: str
-    temperature: int
+    temperature: float
     description: str
